+++ lib/src/instagram.dart
@@ -1,36 +1,50 @@
-import 'dart:ui';
+import 'palette_rgb.dart';
 
-import 'palette.dart';
+/// The official Instagram's brand colors
 
-/// The new Instagram logo has a pleasing palette of reds, oranges, yellows,
-/// blues and violets.
+/// The colors are:
+/// - Royal Blue;
+/// - Blue
+/// - Purple;
+/// - Dark Pink;
+/// - Purple-Red;
+/// - Red;
+/// - Dark Orange;
+/// - Orange;
+/// - Yellow;
+/// - Light Yellow.
 ///
 /// See also:
 ///
-/// - [DesignPieces](https://www.designpieces.com/palette/instagram-new-logo-2016-color-palette)
-class Instagram extends Palette {
-  /// Ctor. Instagram-ish blue (#4c68d7)
-  Instagram.blue();
+/// - [USBrandColors](https://usbrandcolors.com/instagram-colors/)
+class Instagram extends PaletteRGB {
+  /// Ctor. Instagram-ish royal blue (#405DE6)
+  Instagram.royalBlue() : super.opaque(0x405DE6);
 
-  /// Ctor. Instagram-ish marron (#cd486b)
-  Instagram.maroon();
+  /// Ctor. Instagram-ish blue (#5B51D8)
+  Instagram.blue() : super.opaque(0x5B51D8);
 
-  /// Ctor. Instagram-ish orange (#fbad50)
-  Instagram.orange();
+  /// Ctor. Instagram-ish purple (#833AB4)
+  Instagram.purple() : super.opaque(0x833AB4);
 
-  /// Ctor. Instagram-ish purple-violet (#8a3ab9)
-  Instagram.purpleViolet();
+  /// Ctor. Instagram-ish dark pink (#C13584)
+  Instagram.darkPink() : super.opaque(0xC13584);
 
-  /// Ctor. Instagram-ish red-orange (#e95950)
-  Instagram.redOrange();
+  /// Ctor. Instagram-ish purple-red (#E1306C)
+  Instagram.purpleRed() : super.opaque(0xE1306C);
 
-  /// Ctor. Instagram-ish red-violet (#bc2a8d)
-  Instagram.redViolet();
+  /// Ctor. Instagram-ish red (#FD1D1D)
+  Instagram.red() : super.opaque(0xFD1D1D);
 
-  /// Ctor. Instagram-ish yellow (#fccc63)
-  Instagram.yellow();
+  /// Ctor. Instagram-ish dark orange (#F56040)
+  Instagram.darkOrange() : super.opaque(0xF56040);
 
-  /// TODO somehow implement Instagram color.
-  @override
-  Color get color => throw UnimplementedError();
+  /// Ctor. Instagram-ish orange (#F77737)
+  Instagram.orange() : super.opaque(0xF77737);
+
+  /// Ctor. Instagram-ish yellow (#FCAF45)
+  Instagram.yellow() : super.opaque(0xFCAF45);
+
+  /// Ctor. Instagram-ish light yellow (#FFDC80)
+  Instagram.lightYellow() : super.opaque(0xFFDC80);
 }