+++ CHANGELOG.md
@@ -5,5 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Google color palette.
+- Facebook color palette.
+- Whatsapp color palette.
+- Flutter greyish color palette.
 
-* TODO: Describe initial release.
+## [0.0.1] - 2021-04-05
+### Added
+- Instagram color palette [#1](https://github.com/rafamizes/flutter_brand_palettes/issues/1).